+++ NEWS.md
@@ -1,3 +1,5 @@
+# lifecycle (development version)
+
 # lifecycle 1.0.2
 
 * You can now generate arbitrary text in a deprecation message by